feat(cockpit): Repos becomes the fourth tab of Your Throttle
The private per-repository split was its own page and its own rail entry
under System. It is now the fourth tab of Your Throttle, after Overview,
Activity, and Breakdown (owner ask). The rail entry is gone, and /repos
redirects to the page so an existing bookmark still lands.

The old code argued against this in a comment: Repos was "deliberately a
SEPARATE page ... must never ride along on-screen when the throttle is
shown", because which codebases take the owner's time is private, unlike
the shareable voice and surface splits. Folding it in as a tab holds that
line - tabs are mutually exclusive, so Repos still never renders beside
the throttle; the owner has to select it. The "private - only you" badge
moves onto the tab body so the view still announces, on sight, that it is
not for sharing. That stale comment is rewritten rather than left behind
to contradict the code.

Your Throttle already polls GET /stats/data and the repositories ride on
that same envelope, so the tab takes its snapshot as a prop from the
parent instead of opening a second poll of one feed.

Two defects that shipped with the standalone page are fixed here, since
this change moves the code that carried them:

- The four headline cards rendered as an unstyled vertical stack. They
  asked for thr-card-* classes that the Cockpit stylesheet has never
  defined - only the phone's does. They now use the Cockpit's own
  thr-stat tokens, so they match the tiles on the Overview tab.
- The explainer claimed "the amber part of each bar" is the voice share.
  Nothing is amber: the accent is blue and the typed segment is grey. The
  wording now describes what is actually drawn, which matters because
  that section is the honesty note.

The tab list and its stored-value validation move to a pure module and
the validator is derived from the list rather than repeating the keys.
Forgetting to add a new tab there fails quietly - the stored tab is
rejected and the owner is dropped back on Overview on every visit - so it
is now covered by tests.

Cockpit only. The phone's Your Throttle is a single scrolling page with
no tab strip, so its Repos page is untouched.

Verified: typecheck and build clean; 101 cockpit and 430 client-core
tests pass. Driven in a real browser against a stubbed /stats/data - four
tabs present with Repos last, seven repositories ranked, Rank by re-ranks
across turns, characters, and sessions, the tab survives a reload, and
/repos redirects rather than falling into the not-found page.
